Add nonlinear system simulation utility
- Implement `NonlinearSystem` class in `nonlinear_model.py` to
  simulate nonlinear systems.
- Move and rename `utilities/model_simulation.py` to
  `utilities/models/lti_model.py`.
- Add `__init__.py` to `utilities/models` for package
  initialization.
- Update import statements in modules and example scripts to align
  with the new project structure.

This commit adds the `NonlinearSystem` class for simulating
nonlinear functions and improves code organization for simulation
utility modules.
